pyproject.toml
[tool.poetry]
name = "resume-builder"
version = "0.1.0"
description = "YAML interface to easily create and manipulate preset LaTeX format"
authors = ["Aaren Rice <aaren.rice@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
include = '\.pyi?$'
line-length = 119
target-version = ['py311']
